Refactor the main script.
