--- pyfr/scripts/pyfr
@@ -1,1 +1,0 @@
-main.py
